build: rename the sample project to `MyFibonacci`
